Refactor module namespaces
- Use a clean i3pyblocks/modules/__init__.py file
- Move LocalTime module to i3pyblocks.modules.time
- Split tests/test_modules.py in multiple files
